add --verbose flag to indicate INFO logging level
This is useful for fabfiles that use `logging.info` as an intermediate
between 'no output' and 'full debug'.

This is not ideal: I don't like that if/elif check to set the log
level. But I cannot figure out how to do the equivalent of this
argparse snippet:

    parser.add_argument('--verbose', '-v', dest='log_level',
                        action='store_const', const='info', default='warning')
    parser.add_argument('--debug', '-d', dest='log_level',
                        action='store_const', const='debug', default='warning')

I suspect I might be missing some invoke.yml configuration bits as well.

See: #15